Move Log and Time to shared
