+++ casper/src/main/scala/coop/rchain/casper/util/comm/CommUtil.scala
@@ -13,6 +13,7 @@ import coop.rchain.p2p.effects._
 import coop.rchain.comm.transport._, CommMessages._
 import coop.rchain.comm.discovery._
 import coop.rchain.p2p.Network.ErrorHandler
+import coop.rchain.shared._
 
 import scala.util.Try
