+++ comm/src/main/scala/coop/rchain/comm/discovery/KademliaNodeDiscovery.scala
@@ -9,6 +9,7 @@ import cats._, cats.data._, cats.implicits._
 import coop.rchain.catscontrib._, Catscontrib._, ski._, TaskContrib._
 import coop.rchain.comm.transport._, CommunicationResponse._
 import coop.rchain.comm.discovery._
+import coop.rchain.shared._
 
 class TLNodeDiscovery[F[_]: Monad: Capture: Log: Time: Metrics: TransportLayer: Ping](src: PeerNode)
     extends NodeDiscovery[F] {
