+++ comm/src/test/scala/coop/rchain/p2p/EffectsTestInstances.scala
@@ -12,6 +12,7 @@ import coop.rchain.comm._
 import coop.rchain.p2p.effects._
 import coop.rchain.comm.transport._
 import coop.rchain.comm.discovery._
+import coop.rchain.shared._
 
 /** Eagerly evaluated instances to do reasoning about applied effects */
 object EffectsTestInstances {
